[16.0][MIG] - sale_stock_picking_blocking
