+++ sale_stock_picking_blocking/__manifest__.py
@@ -4,9 +4,9 @@
 
 {
     "name": "Sale Stock Picking Blocking",
-    "summary": "Allow you to block the creation of deliveries " "from a sale order.",
-    "version": "15.0.1.0.1",
-    "author": "Eficent, " "Odoo Community Association (OCA)",
+    "summary": "Allow you to block the creation of deliveries from a sale order.",
+    "version": "16.0.1.0.0",
+    "author": "Eficent, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/sale-workflow",
     "category": "Sales",
     "depends": ["sale_stock"],
