+++ sale_stock_picking_blocking/models/__init__.py
@@ -1,5 +1,6 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import sale_order
+from . import sale_order_line
 from . import sale_stock_picking_blocking_reason
 from . import res_partner
